+++ runway/embedded/stacker/stack.py
@@ -62,7 +62,9 @@ class Stack(object):
     def __init__(self, definition, context, variables=None, mappings=None,
                  locked=False, force=False, enabled=True, protected=False):
         self.name = definition.name
-        self.fqn = context.get_fqn(self.name)
+        self.fqn = context.get_fqn(definition.stack_name or self.name)
+        self.region = definition.region
+        self.profile = definition.profile
         self.definition = definition
         self.variables = _gather_variables(definition)
         self.mappings = mappings
@@ -71,14 +73,14 @@ class Stack(object):
         self.enabled = enabled
         self.protected = protected
         self.context = copy.deepcopy(context)
+        self.outputs = None
 
     def __repr__(self):
         return self.fqn
 
     @property
     def requires(self):
-        requires = set([self.context.get_fqn(r) for r in
-                        self.definition.requires or []])
+        requires = set(self.definition.requires or [])
 
         # Add any dependencies based on output lookups
         for variable in self.variables:
@@ -96,10 +98,19 @@ class Stack(object):
                             "within lookup: %s"
                         ) % (variable.name, self.name, lookup.raw)
                         raise ValueError(message)
-                    stack_fqn = self.context.get_fqn(d.stack_name)
-                    requires.add(stack_fqn)
+                    requires.add(d.stack_name)
 
         return requires
+
+    @property
+    def stack_policy(self):
+        if not hasattr(self, "_stack_policy"):
+            self._stack_policy = None
+            if self.definition.stack_policy_path:
+                with open(self.definition.stack_policy_path) as f:
+                    self._stack_policy = f.read()
+
+        return self._stack_policy
 
     @property
     def blueprint(self):
@@ -174,3 +185,6 @@ class Stack(object):
         """
         resolve_variables(self.variables, context, provider)
         self.blueprint.resolve_variables(self.variables)
+
+    def set_outputs(self, outputs):
+        self.outputs = outputs